Disable listener methods if calling them fails
Update issue 832
Status: Done
Labels: bwic
We decided to fix this so that if calling a listener method fails it is
automatically disabled. This prevents the recursion from occurring when
`message` fails and also causes less noise when any other message fails
every time it is called.

This is slightly backwards incompatible change because earlier failures
e.g. in `start_suite` have been ignored (except for the error message)
but now this method is not called anymore after a failure. We agreed that
it is better that libraries take care of the error handling themselves
and this kind of a minor backwards incompatibility is fine. Needs to be
mentioned in the release notes anyway and thus the bwic label.