Keyword type to start/end_keyword listener methods
Update issue 859
Status: Started
Owner: pekka.klarck
Implementation and tests done. Attributes to start_keyword and end_keyword now contain 'type' that is either keyword, suite setup, suite teardown, test setup, or test teardown.

This is now implemented so that if a user keyword is used as a setup or teardown, only the top level keyword gets 'setup/teardown' type and others get type 'keyword'. Is that ok? At least the implementation this way is easier.

This also needs to be documented in the User Guide.