Convert Python docs to Unicode assuming docs are in UTF-8.
Update issue 894
Status: Started
Owner: pekka.klarck
Labels: Target-2.6
Good catch! The problem was actually on Robot's side -- the docs read from Python objects weren't converted to Unicode. I fixed that and now libdoc.py works correctly. Notice that it already worked correctly if resource files contained non-ASCII characters.

A limitation of the fix is that it assumes docs to be in UTF-8 format (unless they already are Unicode). The reason is that there doesn't seem to be any way to dynamically check what encoding a Python modules uses. I still need to document that to the User Guide.

A workaround for your problem, that ought to work also with previous versions, is using Unicode in docs. For example like this:

def example():
  u"""\u4e2d\u6587"""