Don't start XML if suite is not run at all.
Update issue 923
Status: Done
Owner: pekka.klarck
The problem was that Output object was created before creating the
executed TestSuite object. With versions prior to 2.6 this didn't
matter because Output opened the actual XML output file only when it
was used, but in 2.6 the output file is opened when Output object is
created.

The fix was as easy as changing the order in which TestSuite and
Output objects are created. The commit also includes enhancement to
related tests to catch the bug.