Prevent infinite recursion by having a hard limit of how many keywords can be open.
Update issue 551
Status: Done
Labels: Priority-Medium

The infinite recursion problem is now fixed and also acceptance
tested. There is a hard limit of 100 keywords that can be open
simultaneously. That ought to be enough for all valid usages so I
don't consider the fix backwards-incompatible. I also decided to raise
the priority of this issue as crashes are always somewhat nasty and
this has happened several times in real life.

Notice that before the simple fix prevented recursion involving Run
Keyword variants, a separate refactoring was needed to make them use
the same execution context as other keywords.