Renamed robot/runner.py -> robot/run.py.
Also updated the usage accordingly and did some clean-up to it.

Update issue 1044
Status: Done
Now also runner.py has been renamed to run.py. The command line usage
is updated but User Guide and other installation instructions are not.
That will be done as part of documenting the new installation and
start-up system (issue 480) in general.