Implementation and tests for creating variable files as Python or Java classes.
Update issue 1021
Implementation and tests done. The main remaining task is documenting this in
the User Guide but review would be nice too.

The only surprise that I encountered was that static variable files
implemented as Java classes automatically created variables like ${toString}.
The logic we use to ignore such methods when creating Java libs is
somewhat complicated so I decided to use a simpler solution and ignore
all attributes in instances that are callable. That ought to be good enough
solution but needs to be documented.