Fixed BuiltIn.get_variables() to return a dictionary (actually NormalizedDict) instead of _VariableScopes.
Update issue 1180
Status: Done
Labels: bwic
Fixed this by returning variables as a custom NormalizedDict object. It is a subclass of Python's UserDict and thus supports all dictionary methods. A benefit compared to normal dictionaries is that it allows accepting variables in case, space, and underscore insensitive manner similarly as in the test data.

Also noticed that this change is potentially backwards-incompatible, because it is possible that someone has used the returned _VariableScopes on purpose. The change thus needs to be mentioned in the release notes.