Sorting of statistics in log/report.
Update issue 1199
Statistics are now sortable. There is currently no way to restore
original sorting of tags and suites other than refressing the page. I
try to fix that next.

As part of this issue I did some tuning for sorting indicators (those small
arrows) that we already used in report. They are now closer to the right side
of the column and thus take less space. Also the default icon is gray so that
it doesn't distract text close to it too much.

To see how sorting looks like, and to try it out, you can install the latest
version from the version control system or see this example:
http://robotframework.googlecode.com/hg/src/robot/htmldata/rebot/log.html