Documented new `ANSI` value for `--MonitorColors` option.
Update issue 1229
Colors worked on our Windows CI like a charm. The new feature is now also
documented. Acceptance tests still needed before this issue can be closed.

The steps to get colors working with Jenkins on Windows are:

1) Install https://wiki.jenkins-ci.org/display/JENKINS/AnsiColor+Plugin
2) Enable ANSI colors for your build.
3) Run tests with `--MonitorColors ANSI` option.

Above obviously will only work when this feature is distributed with RF 2.7.5.
On other operating system the steps work already now but instead of `ANSI`
you need to use `ON` value with `--MonitorColors`.