Support for returning time in UTC for utils.parse_time and thus to Get Time.
Update issue 1233
Status: Started
Owner: pekka.klarck
Applied the latest patch by aalto.tatu as-is. I'll do some
cleanup/refactoring to it in subsequent commits.

A general problem with the latest patch was that it contained multiple
changesets which made reviewing the patch alone pretty much
impossible. I was able to apply it cleanly and see the changes
afterwards, though, so this wasn't such a big problem. In the future,
please try to make a patch containing everything in one changeset to
make reviewing the patch easier.