Revert the change to round epoch seconds to the closes second instead of down.
Update issue 1233
It turned out that rounding seconds to the closest one and not down,
the change I did while refactoring the code, was not a good idea. The
problem was that you could get epoch seconds in the future. For
example, some of our acceptance tests failed because the following
wasn't guaranteed to be true anymore:

${start} =    Get Time
Create File    ${path}
${modified} =      Get Modified Time  ${path}
Should Be True    '${modified}' >= '${start}'