Testdoc: Fixed rendering output.
Update issue 1286
Main reason testdoc didn't work was that it used createGenerated method from rebot/view.js that had been removed. That was fixed by copying the old functionality into testdoc itself.

A benefit of this change is that testdoc doen't use rebot/view.js anymore. This kind of breakages are thus less likely in the future.

I don't close this issue yet because it seems something is wrong also with testdoc's styles.