RestReader: rafactored the code and fixed handlig relative imports.
Update issue 1495
I started a quite big refactoring to make the code more compatible with
PEP-8 and other code in Robot Framework code base. Also wanted to
separate registering directives from parsing etc. The refactoring is not
yet ready, but because I don't have time to finish it this weekend I
wanted to commit and push what I have done.

The most important reason because I wanted to get the code pushed was
that it contains a temporary fix for handling relative imports in reST
files. This wasn't tested by any parsing related acceptance tests, but a
test for running the Quick Start Guide in reST format fails. The commit
contains also relative imports of empty files in the parsing tests.