Implemented multiple patterns for Remove String Using Regexp. Modified docs.
update issue 1491

It's useful to remove also multiple regexp matches at once. Modified the docs
for Remove String (Using Regexp) and Replace String (Using Regexp) to runnable
format. In original examples the variable to be modified was not set which
made it hard to figure out how the keywords actually modify the input string.