Log: Fixed back button breakage (#1806)
Investigation revealed that resetting hash was only needed if there
were errors in order to scoll back to the correct location. Depending
on browser and number of errors (e.g. with IE8 when there were many
errors), added errors could move the original scroll location.

Nowadays hash is reset only if there are errors after all errors have
been added. Resetting is done using `location.replace()` so no extra
entry is added to history even in that case.

Now that scrollToHash doesn't anymore scoll but only highlights, it
was renamed to highlightLinkTarget.

Tested using IE8 and IE11 on Win7, as well as with modern Firefox and
Chrome on Linux and Win7. At least Safar/OSX still to be tested.