Log: Update element toggle (+/-) icons (#1808).
Took Open Iconic icons into use. Also changed layout so that styling
icons is easier.

Layout changes allowed highlighting icon with yellow when toggle is
active. To make it possible that hovering over the right header
disables toggle highlight (it isn't active then), toggle icon needed
to be placed after the right header in the source. That allowed the
latter of these selectors:

  .element-header:hover > .element-header-toggle {
      background-color: yellow;
  }
  .element-header-right:hover ~ .element-header-toggle {
      background-color: transparent;
  }

Remaining task related to the toggle icons is creating PNG versions to
IE8. @jussimalinen, could you take care of that?