Workaround for logging bug in Python < 2.7.
At program exit logging module calls flush for streams it has
intercepted without looking are they already closed. In Python 2.7
resulting ValueError is silenced, but the error and traceback gets
logged to the console with earlier versions.

We avoid the error by replacing flush method of StringIOs we've used
with a no-op lambda. Earlier the error was silenced because we set
logging.raiseExceptions to False, but nowadays such logging config is
disabled after tests are run. This finally fixes #1821.

Also added an unrelated comment to not-so-obvious code.