Preserve keyword/message order in outputs generated by Rebot.
Model stores keywords and messages a keyword has separately. Now
keywords and messages get a sort key that can be used to sort them
later, and that is used by new `children` property added to keywords.

A cleaner solution would have been storing keywords and messages in
same list, and turn existing `messages` and `keywords` lists to
properties that pick items the common list. That would have required
considerable bigger changes to the model and was thus not done
now. That's something to consider later.

Fixes #1867.