Resolve variables and escapes in suite doc on console.
The problem was that doc was read from the original suite model
object, not from the result object which has doc resolved. Most
information related to the executed suite should be read from the
result object, but in `start_suite` listener methods info about
tests/suites to be run needs to be got from the original model. This
required small enhancements to `ModelCombiner`.

Commit includes tests both for suite doc and for test doc that was
alraedy handled correctly, as well as little cleanup to related test
code.

Fixes #1884.