Support &{dict} variables in variable table
&{dict} variables can now be created in variable table like

    *** Variables ***
    &{DICT}    key=value    foo=${42}

and used in test data as scalars like ${dict}. This is the first major
step in full &{dict} variable support. The next step is making it
possible to use

    Keyword    &{dict}

and also

    *** Variables ***
    &{D}    foo=bar
    &{E}    &{D}    foo=${42}

These two enhancements are covered by #1450. User keywords being able
to accept dict variables similarly as **kwargs in Python is #1561.