UG: Added 'Listener interface version' section back.
It was removed as part of removing references to old RF versions
(#1897). I decided to keep it because although the section is about an
old change, it is still relevant because listener version still needs
to be specified with new style listeners. We are finally going to
deprecate the old version in RF 2.9 (#1841), and this documentation is
going to change then.

Also fixed indentation of the listener methods table.