Made tests that fail on Jython non-critical.
On my machine these fail both with Jython 2.5.3 and 2.7b4. They
apparently pass on CI with Jython 2.5, but I still decided to just
make them non-critical. The main reason is that `OperatingSystem.Start
Process` is already deprecated, and although `OperatingSystem.Run` is
not, running processes on background with it isn't that relevan
functionality.