Support list of instances at ROBOT_LIBRARY_LISTENER
When importing a user library the ROBOT_LIBRARY_LISTENER attribute is
used to register an instance (usually the library itself) as a listener.
It is now possible to either set this attribute to a single instance,
or a list of instances. If a list is provided, all elements will be
registered as listeners.