Libdoc: Show tags and allow viewing only kws with certain tag.
Basically added Tags section that is similar to existing
Shortcuts. Clicking a tag in the Tags section shows only tests with
that tag. Unlike with search, only tests with tags matching fully are
shown (e.g. 'tag' matches 'tag' or 'TAG' but not 'another tag').

The view side of showing tags and searching by them in Libdoc outputs
(#1840) ought to be done now. Some changes still needed to the logic
to get tags embedded to kw docs before the whole issue is done.