Cleanup the jython interpreter after use. #2088
This fix is not very pretty. I would prefer to create an exection context that would cleanup the interpreter and not allow the interpreter to exist after execution stops. But the problem is that RunnerFactory is part of public API. I dont really know if we can change that in a minor release.