Decided not to change how `Set Test/Suite/Global Variable` works.
Earlier we planned to support setting scalar variables so that possible
multiple values would be catenated together similarly as in the Variable
table. That would have required handling non-string values somehow separately,
though, and that was not considered worth the effort. It is anyway easy
to catenate possible long values together using `Catenate` separately.

See #1919 for details.