Write timeouts to output.xml as elements, not attributes.
Earlier timeout was written as an attribute to started test or kw. In
the future we want to be able to set kw timeout based on kw arguments,
and thus timeout cannot be written to output.xml when the keyword
itself starts.

This change is backwards incompatible both because timeout is not
anymore available as an attribute and because we introduced a new
element. We expect timeout to be so rarely needed that possibly
missing it isn't a big issue and also that output.xml consumers do not
break if they encounter a new element.