Issue #2245:Supporting time for "Repeat Keyword"
"Repeat Keyword" repeats the keyword specified times basically.
This commit is to support time string (in robot's time format) as the
the first argument of the "Repeat Keyword".
When a time string passed to the first argument, "Repeat Keyword"
repeats the keyword for the specified time regardless of how many times
it repeats.
A number like 42 is interpreted as a count instead of 42 seconds for
backward compatibility. Also floating point numbers like 1.5 would
raise exception just like before for same reason.