Listener v3 implementation. #1208
Implementation on the listener API ought to be pretty much done.
Test listener included but it is not used for actual testing.
Docs missing too.

Also added `insert` method to ItemList. Thought it would be needed by
the new API or its unit tests but it seems it isn't. It's generally
useful so decided to leave it.