Small cleanup and added utest for handling settings between runs.
PR #2438 by @krizex already fixed the problem with settings preserving
some date between runds he had reported as #2437. This commit slightly
cleans up the code and unit tests in the PR. A new unit/integration
test to validate the real problem with --rerunfailed that this problem
caused is added too.