Libdoc: Syntax highlighting when using reST. #2448
Added Pygments styles and configured docutils to be compatible with
them.  Embedding Pygments styles to all Libdoc outputs may be a bit
questionable, but they are small enough that I don't think it is a
real problem. We may also later add syntax highlighting to Libdoc's
default syntax too.

Also made background color used by examples etc. in Libdoc outputs and
also in logs and reports a little lighter. Syntax highlighting looks
better on a little lighter background and this background color is
used also in the User Guide.