Avoid importing `String/DateTime` when importing `string/datetime`
`import string` and `import datetime` should always import Python
standard modules, but in some cases on Windows (and possibly OSX) they
imported Robot's `String` and `DateTime`. This occurred if
`PYTHONCASEOK` was set, but also sometimes if installation was on a
directory shared between a virtual machine and a host machine.

After the fix relative `String.py` and `DateTime.py` won't be imported
anymore. On Python 3 they were never imported.

Fixes #1433.