Introduce an option for Tidy to set the max line length
Tidy currently reformats suites with eight cells per line regardless of
line length. This enhancement adds a `--maxlinelength` argument to tidy
that splits the line at the specified line length when possible.

A summary of changes introduced in this commit:

* Introduced a `--maxlinelength` option for Tidy
* `--maxlinelength` defaults to 80 characters by default
* It is only applicable to space separated and pipe separated robot
files. Tabs can have a configurable length, and many 3rd party HTML
formatters exist already which can do a very good job.
* Updated existing test data to have the correct line default length
(80 characters)
* Added new tests and new test data to test custom line length
* Updated usage text for Tidy with information about the option
* Update the userguide
* This PR however introduces a tighter coupling between rowsplitter.py,
formatters.py and filewriters.py. It may not be desirable.

Changes to the rowsplitter logic:
Originally, the row splitter would either split the row at 8 cells or split
depending on the presence of IF, ELSE, etc. so that they begin on a new
line.

Now, it calculates the line length based on how the row and the table
would be formatted, and calculates split indices based on that. It would
always split at 80 characters by default. And it only use the new login for
space and pipe separated txt files.