Access variable items using `${var}[item]` syntax.
This works if `${var}` is a list or dictionary and should be used
instead of the old `@{var}[item]` and `&{var}[item]` syntax that will
later be deprecated and changed. See #2601.

The change is backwards-incompatible, but possible problems are easy
to fix (like done in this commit). The change still needs to be
documented.