Use "skipped" instead of "skip" in testsuite
This is due to xunit jenkins plugin not supporting "skip" attribute
anymore. If you are using tools such as jenkins, this change will be
backwards compatible but if you have some scripting that will try to
read the "skip" attribute from  testsuite, you need to change that into
"skipped".

Addresses: #2292
Fixes: #2978