Disable type conversion with `@keyword(types=None)`.
Makes it easier to disable type conversion altogether, including based
on default values, when it is not desired. For example, eases fixing
regression in Telnet caused by automatic conversion when `None` is a
default value.

Change is related to issue #2947 (specifying types using the
`@keyword` decorator) and to issue #2932 (getting types from default
values).