3.1b1 release notes: `${var}[xxx]` bwic issue shown on top level
Issue #2601 introduced `${var}[xxx]` syntax to access variable items.
We got #2996 and #3004 about problems using this syntax to that the
part in the square brackets should be considered literal. It is easy
to use `${var}\[xxx]` to accomplish that, but this possible problem
needs to be shown more prominently in release notes.