Fix regression with --suite.
Attempts to always parse only files matching --suite introduced by
issue #2564 and PR #2583 caused severe regressions due to files that
should have been parsed being ignored. The problems are explained more
thoroughly in issue #3029.

The fix is to return back to functionality in RF < 3.1 and only ignore
files based on their name, not by "long name". This is not 100%
correct and some files are parsed unnecessarily, but that doesn't
really matter because parsing is fast and tests are ignored from
execution correctly later. This was already discussed in the comments
of #2564 when I explained why I didn't consider the issue that high
priority.

Fixes #3029.