Inline Python evaluation support using `${{expression}}` syntax
Implementation and tests ready for basic functionality. Still need to
decide do we want to add some modules to the evaluation namespace
automatically and/or do we want allow importing additional modules.
See the discussion in the issue #3179 for more details about that and
about this issue in general.

Documentation is also totally missing.