xmllogger: Allow for addition of timezone and daylight savings to status
The timestamp is localtime, so if output files are moved across time
borders, they become less valuable then they could be.
If this is important to someone, then add the ability to save a timezone
and dst attribute in the status output of xml files.
e.g.
<status status="FAIL" starttime="20200909 12:38:31.781"
endtime="20200909 12:38:31.788" timezone="10.0" dst="0"></status>