robottime: Allow for utc timestamps.
Use module variables to specify an 'override' for the timestamp into
utc.
User must change use_utc_timestamp module variable for utc to be used.
Allowing for this as current timestamp is in local time, so use in
databases may result in invalid date time search results if not set to
same timezone.