robottime: Allow for iso8601 utc timestamps.
Use module variables to specify an 'override' for the timestamp into
utc and or iso8601 format.
User must call default_to_iso8601() if they wish to use it, prior to
doing an timestamp functionality.
Borrow parsing routines from https://github.com/salabs/TestArchiver.
Allowing for this as current timestamp is in local time without timezone
informatoin, so it is unclear in output files when exactly the test was
run.