settings: Add timestampformat setting/command line argument.
To prepare for allowing iso8601utc option to be selected by the user add
a commandline options or setting for it.
Initially we will only allow iso8601utc as a value here, but theoretically
this could become an arbitrary time format string in time. That would
require a much larger rewrite of utils/robottime though.