timestamp: Adjust timestamp used when timestampformat specified in options
If the user specifies iso8601utc as the timestampformat update the
robottime module by calling default_to_utc_timestamp() and
default_to_iso8601() which will globally adjust the module to output
timestamps in iso8601 utc with Z as suffix.
e.g. 2020-09-10T06:59:11.436Z
The intent is that in the future arbitrary formats could be specified.