time: Raise ValueError rather than blanket Exception on timestamp parse
Raising the base exception meant that higher level functions which were
passing on ValueError failed incorrectly. Changing to throw a ValueError
ensure alternate parse methods are used.

e.g. on Get Time keyword, NOW is used and results in a normalised time
of 'NOW00000 00:00:00.000', which results in an exception when parsing,
hence acceptance and unit tests were failing.