Move None conversion fully to NoneConverter.
This makes conversion a bit more strict as now explicitly specified
NoneType conversion fails if the value isn't NONE (case-insensitively).
I doupt that makes any difference now because using NoneType as the
only type sounds pretty strange, but it's likely an important feature
when we add support for Union.

This is related to #3729.