Make Should Be Title compatible Convert To Title
'Convert to Title Case' was added in Robot Framework 3.2. 'Should be
Title Case' was found not to be compatible with the new 'Convert to
Title Case' keyword. The new 'Convert to Title Case' utilized a
custom algorithm to convert a string to title case. For example, words
like 'OK' and 'iPhone' are considered to be in title case.

'Should be Title Case' was using str.istitle() as the internal function
to determine if a string was in title case. This function does NOT
consider 'OK' or 'iPhone' to be title case.

This discrepancy resulted in some strings converted to title case were
not passing "Should Be Title Case".

To fix this, 'Should Be Title Case' now uses the 'convert_to_title_case'
function internally to compare with the input string to determine if
it is in title case.