Update deprecation warning related to disabling arg resolving.
The current system is hard to use and we have #2190 about enhancing
it. The change didn't happen in RF 3.1 like the deprecation warning
stated and changed that to "in the future".

This is useful functionality, needed right now by #3648, so I hope we
really get that mess sorted out some day.