Enhance error reporting with IF/ELSE with invalid condition.
IF, ELSE IF and ELSE are now all lexed regardless now many conditions
they have. Parser then validates that they have correct conditions
(i.e. IF and ELSE IF have one and ELSE has none).

Added comments that it would be better to mark these (as well as same
situation with END) errors already by lexer. Doing that will require
parser changes and it can be done later.

Also little fine-tuning for IF/ELSE error messages in general as well
as some class and attribute renaming.