IF refactoring.
IF model is enhanced so that instead of ELSE IF and ELSE being in IF's
`body`, IF has `orelse` that contains the first ELSE IF or ELSE, which
can then contain the next one in its `orelse`, and so on. This is how
Python stores if/elif/else when creating AST.

Parsing needed to be enhanced a bit to support `orelse`, but it makes
further processing of IF structures, especially validation and
execution, a lot simpler. An added benefit is that now ELSE IF and
ELSE have their own nodes in the AST and thus their line numbers are
correct in listeners. Tests for IF with listeners need to be added
still, though.

Additional changes:
- Renamed some IF related model objects to be consistent with FOR.
- Tests for running and skipping IF/ELSE IF/ELSE branches.
- Unit tests for IF AST.