model: change .keywords -> .body
This is continuation to previous change where .setup and .teardown were
removed from the .keywords. On the whole, these names are more proper,
since the .body contains both keywords and control statements like for
and if.

The .keywords property is left in place as a deprecated read-only
attribute. Any modifications attempts through .keywords will raise an
AttributeError.