Better separation of "active" and "non-active" keyword.
Non-active keywords are needed with setup/teardown when the parent
object don't have them set.

Earlier keyword was non-active (i.e. was considered false) when name
was an empty string. This was problematic because that's the default
value for name and especially unit tests often don't give explicit
name. Now keyword is considered false/non-active only if name is None.