Refactor parsing output.xml.
Main motivation was not creating child handler instances when a
handler is created to avoid problems with cyclic handlers. This
allows, for example, separating keyword and IF parsing. Earlier
keyword handler needing if handler and if handler needing keyword
handler made that impossible. This is to some extend related to
model refactoring discussed in #3749.