Store IF in output.xml as nested structure, not as sequence.
This was considered already earlier because the model is represented
in output.xml better this way. The main motivation to do it now was
that output.xml generated by Rebot alrady stored IFs in this manner
and changing that didnt' make sense.

This commit also introduces new `branch_status` attribute to If
objects. It is needed because `status` represents the whole stucture
but individual branches may also be `NOT_RUN`.