Try to fix unit test failure on CI.
These tests fail on CI like this:

```
Traceback (most recent call last):
  File "/home/runner/work/robotframework/robotframework/utest/output/test_filelogger.py", line 23, in setUp
    robottime.TIMESTAMP_CACHE = _FakeTimeCache()
  File "/home/runner/work/robotframework/robotframework/utest/output/test_filelogger.py", line 13, in __init__
    self.fake = time.mktime((2006, 6, 13, 8, 37, 42, 0, 0, 1)) + 0.123
OverflowError: mktime argument out of range
```

It's documented that `time.mktime` can fail like that if the underlying
OS doesn't support dates to so distant past. It sounds super strange
that yeat 2006 wouldn't be supported, but this is such a strange error
that using 2021 is worth the effort.