Add HTML support to exceptions exposed via robot.api.
Exposing these exceptions was proposed by #3685 and this enhancement
can be considered to be part of that.

Fixes #3866 that explicitly asked to add HTML support to these
excpetions.

A functional change caused by this change is that these exceptions now
require a message. That means, for example, that

    raise SkipExecution

doesn't work anymore and you need to use

    raise SkipExecution('Some message')

instead. I consider that good but if someone has use cases for raising
these exceptions without a message changing this can be still changed.