Refactor: Rename CombinedConverter.{args,types}
A bit questionable to rename a public attribute in a bug fix release,
but this class itself is only for internal usage and also brand new so
nobody ought to use it. Possible users neeed to use `hasattr` or
`getattr` to handle the differences. If someone reports using this
before RF 4.0.1, I promise to add `args` back as a deprecated property.