Fix using IntEnum in type conversion.
Earlier IntEnum based type hints were considered to be integers, not
enums. They are kind of both, but in our case we should handle them
as enums.

This was fixed by moving EnumConverter creation before IntConverter.
It works because converters are registered atthe same time and the
first registered converter has presedence. Now that we have this many
converters the ordering being so important isn't too great. Perhaps we
should make priority explicit when registering converters instead.

This is the first and more important part of fixing #3910.