Avoid accessing `Keyword.teardown` when keyword has no teardown.
Accessing `Keyword.teardown` creates another `Keyword` object that
represents the teardown. Its done even if the keyword has no teardown
to make the API easier to use. It's convenient to be able to use e.g.

    kw.teardown.name = 'Example'

instead of

    if kw.teardown is not None:
        kw.teardown.name = 'Example'

The problem is that we in few cases want to do something with the
teardown if it exists and just using

    if kw.teardown:
        ...

is enough to create a `Keyword` representing the teardown.

This commit introduces new `Keyword.has_teardown` property that can be
used like

    if kw.has_teardown:
        ...

without creating a new `Keyword` object.

This saves surprisingly big amount of memory. See #4114 for more
information.