Rename utils.unic to utils.safe_str.
"unic" is a bit strange name with Python 3 when strings are Unicode.

The original reason to add this util was to avoid problems when using
byte-strings and Unicode-strings together with Python 2. That problem
we don't have anymore with Python 3, but the util still handles
strange (one could say broken) objects that cannot be converted to a
string. It also formats bytes differently than Python 3 and handles
NFC normalization. Byte formatting is perhaps a big questionable and
NFC normalization should probably be provided an independent util.

Part of this change is avoiding using this util altogether where just
str is safe.