Enhance logging unexecuted WHILE and FOR loops.
WHILE loops (#4084) having condition already initially False will get
NOT RUN status. This was how WHILE worked earlier, but it was changed
recently when fixing problems with invalid conditions.

FOR loops that have nothing to iterate over will also get NOT RUN
status. In addition to that, their body is run once, with NOT RUN
status, to show what would have been executed. This fixes #4184.