Don't retry libdoc with :: and show all error messages.
This fixes a case where the user trying to build a library
(say `test_lib`) with arguments such as:

`test_lib::d:/testRepo/functions2.txt`

fails.

Given that `::` is present, it's no longer retried as a resource.

Also, the message shown to the user will now mention all errors
and not just the message saying that the resource file does not exist.