Remove seemingly inefficient optimization.
The optimization was part of PR #4350. This particular change
basically avoided a function call when finding a lexer for a new
statement. It was a bit ugly and on my machine brought no measurable
performance benefits. @fabioz reported 2% enhancement with his tests
but even that's pretty little. Python 3.11 should also make function
calls faster in general:
https://docs.python.org/3.11/whatsnew/3.11.html#faster-cpython