No warning if there is one public and one or more private keywords.
The warning was initially added when the support for private keywords
was added (#430). It was needed because we otherwise favored public
keywords even when the private keyword was in the same resource file.
Now we prefer local keywords over external keywords (#4366) and it's
safe to use imported public keywords instead of imported private
keywords.