Adjust code and tests to new `typing.get_type_hints` semantics.
Python 3.11 changed `typing.get_type_hints` so that `None` isn't
anymore added to returned types automatically if an argument has
`None` as a default value. That changed our argument conversion logic
so that if you have an argument like `arg: str = None` and pass it a
`None` object, it isn't passed as-is but instead converted to a string
`None`. Argument conversion depending on Python versions wouldn't be
great, so this commit changes the logic so that `None` is never
converted if an argument has `None` as a default.

The same change also changed how types are shown by Libdoc. Earlier
`arg: str = None` was shown as `arg: str | None = None`, but nowadays
it's just `arg: str = None`. I prefer the new formatting and just
updated tests so that they expect that if Python 3.11+.

This is part of Python 3.11 compatibility. #4401