Faster searching of variables in strings.
Based on `timeit` this implementation is considerably faster. For
example,

```
python -m timeit 'from robot.variables.search import search_variable' 'search_variable("${var}")'
```

reports ~3.5usec per loop with the old implemenation and only ~2.0usec
per loop with the new one. Although the difference in percentages is
huge, also the old code was so was that this doesn't affect the actual
execution times too much. For example, the execution time of the
following test dropped from ~910msec to ~880msec. Not a huge drop, but
the difference is measurable and adds up with longer suites.

```
*** Variables ***
${var}     var
@{list}    one    two

*** Test Cases ***
Example
    FOR    ${i}    IN RANGE    1000
       Log    ${var}
       Log    Some text with ${var} and more with ${i + ${1}} still more with ${list}[0]
    END
```

One reason for the speedup is having the core logic in loop without
any function calls. That results with somewhat complicated code, but
the old state machine wasn't that easy to understand either and this
implementation is quite a bit shorter.