Enhance reporting errors if multiple keywords match.
Earlier conflicts with keywords having embedded args were handled when
getting a handler from a particular file. Now all matching handlers
are returned and all conflicts are handled in the same place.

The main motivation for this is making it possible to select the best
match if there are conflicts with embedded args as proposed in #4454.
Earlier implementation allowed handling conflicts only within a single
file, but now we can handle them across all files.