Give precedence to local and public kws over search order.
Prioritizing local keywords in a resource file (#4366) and
prioritizing public user keywords over private ones (#430) are new
features in RF 5.1. Earlier possible library search order was looked
at first, but this commit changes logic so that it is consulted only
as the last resort. Motivation for the change has been discussed in
the comments of #4454:

https://github.com/robotframework/robotframework/issues/4454#issuecomment-1248018031