Make DataError normal catchable error by default.
It can be turned into a syntax error with `syntax=True`.

Makes IF and WHILE condition errors not being catchable.
That fixes #4348.

Also changes error if a keyword is used with wrong number of
arguments to a normal error. That's consistent with keywords
not being found already earlier causing catchable errors.

There could be similar semantical changes already elsewhere.