Always support Boolean conversion with 'True/False' strings.
Now they are supported even if English would be disabled (which isn't
yet possible during execution). They were left to the En class as
examples that other language classes can translate.

Notice that although 'None' is listed here, it's not actually used for
anything at least now. The reason is that Boolean conversion expicitly
checks is the value 'None' (case-insenstively) and returns Python
`None` in that case instead of `False`.