Consistent type conversion with `arg: T = None`.
Prior to Python 3.11 that syntax was considered same as
`arg: T|None = None` and with unions we don't look at the default
value at all if `T` isn't known.

This commit makes behavior with Python 3.11 consistent with how
conversion works with older Python versions. Fixes #4626.

Might be better not to look at the default values in general if an
argument has type information. That would be a backwards incompatible
change, though, and needs to wait for a major version.