Consistent handling of unrecognized types in union conversion.
After this change `arg: int|Unrecognized` and `arg: Unrecognized|int`
behave the same way.  Earlier `int` conversions wasn't attempted in
the latter case because the unrecognized type was encountered first.
Fixes #4648.

After this change `arg: T = None` and `arg: T|None = None` are handled
the same way regardless is `T` known or not. That means that it was
possible to remove the code needed to make `arg: T = None` behave
consistently with different Python versions (#4626).