Libdoc: Show parameterized types properly in HTML #4538
Earlier, for example, `list[int]` was just a single type with a link
to type info about lists. Now `list` is a link to list info and `int`
is a link to integer info.  Specs were enhanced already earlier. This
change uses enhanced specs to show info in HTML.

Also showing unions was changed as part of this so that `|` is used
instead of `or`. The motivation was to make syntax look like Python
both with parameterized types and with unions.

Info is now shown like this:

    a: int | float           # union, ok
    b: list [ int ]          # one arg, perhaps too much spaces
    c: list [ int | float ]  # union arg, also lot of spaces
    d: dict [ str , int ]    # two args, space before comma is odd

There are some arguably unnecessary spaces. Removing them isn't easy
with our current templating system, so it might be best to just leave
them as they are.