Remove zipapp section added as part of #4613.
Creating zipapps is more about distribution than installation. It
could be documented somewhere in the User Guide, but I'm not sure is
that really needed either. There's nothing special about that with
Robot other than the need to use `importlib_resources` with Python 3.9
and older, but we have an explicit error in the code related to that.