Initial custom parser implementation (#1283).
The code works and the code itself and the new API also look pretty
good for me. There is, however, plenty of work still:

- Acceptance tests. My initial tests work well, but they neeed to be
  converted to proper acceptance tests.

- Documentation in source (--help, API docs, ...). I added initial API
  docs but there are some holes and --help text is completely
  missing. I added FIXMEs to places where more docs are needed.

- User Guide documentation.

- `Defaults` API. We pass `Defaults` as the second argument to `parse`
  and `parse_init`, but that class itself doesn't have too good API.
  It needs to be enhanced and documented (incl. types) properly. The
  class could possibly also get a better name and it needs to be
  exposes directly via `robot.running`.

Although the feature is still work-in-progress, I commit it now to
make it possible for others to test the new API. It can the be still
enhanced based on feedback.