Create timestamps during execution using `datetime.now()`.
This is the first step of using `datatime` for timestamps (#4258).
Result objects now get their timestamps using `datetime.now()` and
they are stored to `start_time` and `end_time` attributes. Old
`starttime` and `endtime` are propertyes that return same string
representation as earlier.

Timestamps are still stored to output.xml in the old format. Moving to
ISO 8601 format for performance and standard compatibility is the next
step.