Document that conversion in cases like `arg: type = default` will change.
We currently convert arguments firts based on type hints and then
based on possible default value types. It would be better to do
default value based conversion only if there's no type hint, but that
would make the conversion logic Python version dependent. Better to
wait until we have Python 3.11 as the minimum version, but the planned
change is now documented.

See #4881 for more information (and reopen it if/when the change is
done).