Don't report parsing errors with tests and keywords further.
We only validate that body and name aren't empty, and errors being
reported would mean that tools adding content to body needed to clear
them.

With tests we already validated that body and name aren't empty also
during execution, so with them not reporting these problems as parsing
errors didn't require other changes. With keywords same runtime
validation needed to be added, but that's a good idea in general. This
way programmatically created keywords cannot be empty either.

Fixes #4880.