Improvement for Robot "Get Length" with len() Function own
During my work on API endpoint validation, I needed know the length of an integer variable. It was then that I realized that the "Get Length" function couldn't count the characters in variables of types int, bool, and float. In discussions with my team, we noticed this need in several projects. Sometimes, it's necessary to validate the number of digits in an integer variable to ensure it doesn't exceed the limit of a system's business rule.

I'm submitting a Pull Request with this potential improvement. Now, when using "Get Length," it will return the number of digits in variables of types str, int, bool, and float and will no longer display the message "Could not get length of."