Enhance validation of control structure options.
Most importantly, validate values that have a certain set of accepted
values already in parser when possible. Because values can be set also
as variables, they need to be validated again at the execution time.

Also enhance related error messages in general.

Part of the motivation is making it easy to validate `scope` accepted
by the forthcoming `VAR` syntax (#3761).