Support `separator` option in Variables section.
This option specifies the separator to use if a scalar variable gets
more than one value. The old way to accomplish that is using a
`SEPARATOR` marker as the first value. The configuration option is
already recognized by the parser and also consistent with the
`separator` option the new `VAR` syntax (#3761) will get.

Fixes #4896.